app/src/main/java/com/prueba/pruebatecnica/ui/userlist/UserListViewModel.kt
package com.prueba.pruebatecnica.ui.userlist


import androidx.lifecycle.ViewModel
import androidx.lifecycle.viewModelScope
import com.prueba.pruebatecnica.data.model.User
import com.prueba.pruebatecnica.data.repository.UserRepository
import dagger.hilt.android.lifecycle.HiltViewModel
import kotlinx.coroutines.flow.MutableStateFlow
import kotlinx.coroutines.flow.SharingStarted
import kotlinx.coroutines.flow.StateFlow
import kotlinx.coroutines.flow.stateIn
import kotlinx.coroutines.flow.update
import javax.inject.Inject

@HiltViewModel
class UserListViewModel @Inject constructor(private val repository: UserRepository) : ViewModel() {
    val users: StateFlow<List<User>> = repository.getUsers().stateIn(
        scope = viewModelScope,
        started = SharingStarted.Lazily,
        initialValue = emptyList()
    )

    private val _selectedUser = MutableStateFlow<User?>(null)
    val selectedUser: StateFlow<User?> = _selectedUser

    fun selectUser(user: User) {
        _selectedUser.update { user }
    }
}